+++ entity-view/impl/src/main/java/com/blazebit/persistence/view/impl/mapper/CollectionAddMapper.java
@@ -18,25 +18,27 @@ package com.blazebit.persistence.view.impl.mapper;
 
 import com.blazebit.persistence.view.impl.accessor.AttributeAccessor;
 
+import java.util.Collection;
+
 /**
  *
  * @author Christian Beikov
- * @since 1.2.0
+ * @since 1.3.0
  */
-public class SimpleMapper<S, T> implements Mapper<S, T> {
+public class CollectionAddMapper<S, T> implements Mapper<S, T> {
 
     private final AttributeAccessor attributeAccessor;
 
-    public SimpleMapper(AttributeAccessor attributeAccessor) {
+    public CollectionAddMapper(AttributeAccessor attributeAccessor) {
         this.attributeAccessor = attributeAccessor;
     }
 
     @Override
     public void map(S source, T target) {
-        attributeAccessor.setValue(target, source);
+        Collection<Object> value = (Collection<Object>) attributeAccessor.getValue(target);
+        if (value != null) {
+            value.add(source);
+        }
     }
-
-    public AttributeAccessor getAttributeAccessor() {
-        return attributeAccessor;
-    }
+    
 }